[seller] refactor(228): 상품정보 제공고시  constant를 entity에서 features로 이동 및 파일 통합
